Use correct size_t type instead of unsigned int for the length argument in vars_ functions.

+++ libwzd-core/wzd_vars.c
@@ -70,7 +70,7 @@ static struct wzd_shm_vars_t * _shm_vars[32] = { NULL };
 
 
 
-int vars_get(const char *varname, char *data, unsigned int datalength, wzd_config_t * config)
+int vars_get(const char *varname, char *data, size_t datalength, wzd_config_t * config)
 {
   if (!config) return 1;
 
@@ -140,7 +140,7 @@ int vars_get(const char *varname, char *data, unsigned int datalength, wzd_confi
  *
  * \todo we should change the value in config->cfg_file
  */
-int vars_set(const char *varname, const char *data, unsigned int datalength, wzd_config_t * config)
+int vars_set(const char *varname, const char *data, size_t datalength, wzd_config_t * config)
 {
   int i;
   unsigned long ul;
@@ -214,7 +214,7 @@ int vars_set(const char *varname, const char *data, unsigned int datalength, wzd
   return 1;
 }
 
-int vars_user_get(const char *username, const char *varname, char *data, unsigned int datalength, wzd_config_t * config)
+int vars_user_get(const char *username, const char *varname, char *data, size_t datalength, wzd_config_t * config)
 {
   wzd_user_t * user;
   wzd_group_t * group;
@@ -330,7 +330,7 @@ int vars_user_delip(const char *username, const char *ip, wzd_config_t *config)
   return backend_mod_user(config->backends->filename, user->uid, user, _USER_IP);
 }
 
-int vars_user_set(const char *username, const char *varname, const char *data, unsigned int datalength, wzd_config_t * config)
+int vars_user_set(const char *username, const char *varname, const char *data, size_t datalength, wzd_config_t * config)
 {
   wzd_user_t * user;
   unsigned long mod_type;
@@ -500,7 +500,7 @@ int vars_user_new(const char *username, const char *pass, const char *groupname,
   return err ? 1 : 0;
 }
 
-int vars_group_get(const char *groupname, const char *varname, char *data, unsigned int datalength, wzd_config_t * config)
+int vars_group_get(const char *groupname, const char *varname, char *data, size_t datalength, wzd_config_t * config)
 {
   wzd_group_t * group;
 
@@ -536,7 +536,7 @@ int vars_group_get(const char *groupname, const char *varname, char *data, unsig
   return 1;
 }
 
-int vars_group_set(const char *groupname, const char *varname, const char *data, unsigned int datalength, wzd_config_t * config)
+int vars_group_set(const char *groupname, const char *varname, const char *data, size_t datalength, wzd_config_t * config)
 {
   wzd_group_t * group;
   unsigned long mod_type;
@@ -709,7 +709,7 @@ struct wzd_shm_vars_t * vars_shm_find(const char *varname, wzd_config_t * config
 /* fills data with varname content, max size: datalength
  * @returns 0 if ok, 1 if an error occured
  */
-int vars_shm_get(const char *varname, char *data, unsigned int datalength, wzd_config_t * config)
+int vars_shm_get(const char *varname, char *data, size_t datalength, wzd_config_t * config)
 {
   struct wzd_shm_vars_t * var;
   int ret = 1;
@@ -730,7 +730,7 @@ int vars_shm_get(const char *varname, char *data, unsigned int datalength, wzd_c
  * Create varname if needed.
  * @returns 0 if ok, 1 if an error occured
  */
-int vars_shm_set(const char *varname, const char *data, unsigned int datalength, wzd_config_t * config)
+int vars_shm_set(const char *varname, const char *data, size_t datalength, wzd_config_t * config)
 {
   struct wzd_shm_vars_t * var;
