+++ libwzd-core/wzd_vars.h
@@ -40,7 +40,7 @@
 struct wzd_shm_vars_t {
   char *key;
   void * data;
-  unsigned long datalength;
+  size_t datalength;
 
   struct wzd_shm_vars_t * next_var;
 };
@@ -48,17 +48,17 @@ struct wzd_shm_vars_t {
 /** \brief Get the value of a server variable
  * \returns 0 if ok, 1 if an error occured
  */
-int vars_get(const char *varname, char *data, unsigned int datalength, wzd_config_t * config);
+int vars_get(const char *varname, char *data, size_t datalength, wzd_config_t * config);
 
 /** \brief Change value of server variable
  * \returns 0 if ok, 1 if an error occured
  */
-int vars_set(const char *varname, const char *data, unsigned int datalength, wzd_config_t * config);
+int vars_set(const char *varname, const char *data, size_t datalength, wzd_config_t * config);
 
 /** \brief Get the value of a user variable
  * \returns 0 if ok, 1 if an error occured
  */
-int vars_user_get(const char *username, const char *varname, char *data, unsigned int datalength, wzd_config_t * config);
+int vars_user_get(const char *username, const char *varname, char *data, size_t datalength, wzd_config_t * config);
 
 /** \brief Change value of user variable
  * \param username the name of the user
@@ -68,7 +68,7 @@ int vars_user_get(const char *username, const char *varname, char *data, unsigne
  * \param config the running server config
  * \returns 0 if ok, 1 if an error occured
  */
-int vars_user_set(const char *username, const char *varname, const char *data, unsigned int datalength, wzd_config_t * config);
+int vars_user_set(const char *username, const char *varname, const char *data, size_t datalength, wzd_config_t * config);
 
 /** \brief Create a new user
  * \returns 0 if ok, 1 if an error occured
@@ -91,7 +91,7 @@ int vars_user_delip(const char *username, const char *ip, wzd_config_t *config);
 /** \brief Get the value of a group variable
  * \returns 0 if ok, 1 if an error occured
  */
-int vars_group_get(const char *groupname, const char *varname, char *data, unsigned int datalength, wzd_config_t * config);
+int vars_group_get(const char *groupname, const char *varname, char *data, size_t datalength, wzd_config_t * config);
 
 /** \brief Change value of group variable
  * \param groupname the name of the group
@@ -101,7 +101,7 @@ int vars_group_get(const char *groupname, const char *varname, char *data, unsig
  * \param config the running server config
  * \returns 0 if ok, 1 if an error occured
  */
-int vars_group_set(const char *groupname, const char *varname, const char *data, unsigned int datalength, wzd_config_t * config);
+int vars_group_set(const char *groupname, const char *varname, const char *data, size_t datalength, wzd_config_t * config);
 
 /** create a new group
  * @returns 0 if ok, 1 if an error occured
@@ -120,12 +120,12 @@ struct wzd_shm_vars_t * vars_shm_find(const char *varname, wzd_config_t * config
 /** fills data with varname content, max size: datalength
  * @returns 0 if ok, 1 if an error occured
  */
-int vars_shm_get(const char *varname, char *data, unsigned int datalength, wzd_config_t * config);
+int vars_shm_get(const char *varname, char *data, size_t datalength, wzd_config_t * config);
 
 /** change varname with data contents size of data is datalength
  * @returns 0 if ok, 1 if an error occured
  */
-int vars_shm_set(const char *varname, const char *data, unsigned int datalength, wzd_config_t * config);
+int vars_shm_set(const char *varname, const char *data, size_t datalength, wzd_config_t * config);
 
 /** @} */
